+++ shifter/shifter_platform/mission_control/api/resources.py
@@ -1,0 +1,233 @@
+"""NGFW, credential, and script DRF views for Mission Control."""
+
+from __future__ import annotations
+
+import logging
+import os
+from typing import Any
+
+from django.contrib.auth.models import User
+from django.http import Http404, JsonResponse
+from rest_framework.request import Request
+from rest_framework.response import Response
+
+from cms.services import ScriptUploadError, complete_script_upload, initiate_script_upload, list_scripts
+from cms.services import delete_credential as cms_delete_credential
+from cms.services import list_ngfws as cms_list_ngfws
+from mission_control.api._base import (
+    MissionControlAPIView,
+    MissionControlReadAPIView,
+    _credentials_write_permission,
+    _ngfw_read_permission,
+    _ngfw_write_permission,
+    _script_read_permission,
+    _script_write_permission,
+    _validated,
+)
+from mission_control.api.permissions import HasMissionControlActor
+from mission_control.api.serializers import (
+    CredentialCreateSerializer,
+    NGFWCreateSerializer,
+    NGFWDestroySerializer,
+    ScriptUploadSerializer,
+)
+from mission_control.views._common import _pkg
+from mission_control.views._credentials import _CredentialError, _persist_credential, _validate_credential_spec
+from mission_control.views._ngfw import _extract_ngfw_create_payload, _NgfwError, _run_ngfw_destroy
+from shared.api.permissions import IsAuthenticatedSessionOrApiToken
+from shared.errors import classify_user_message
+from shared.exceptions import CMSError
+from shared.log_sanitize import safe_log_value
+
+logger = logging.getLogger(__name__)
+
+
+class NGFWCreateView(MissionControlAPIView):
+    """Create a new NGFW."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _ngfw_write_permission()]
+
+    def post(self, request: Request) -> Response | JsonResponse:
+        """Start NGFW provisioning for the authenticated actor."""
+        data, error = _validated(self, NGFWCreateSerializer, request.data)
+        if error is not None:
+            return error
+        assert data is not None
+        user = self.actor_user()
+        payload = _extract_ngfw_create_payload(data)
+        try:
+            ngfw_ref = _pkg().cms_create_ngfw(user=user, **payload)
+        except (TypeError, ValueError, CMSError) as exc:
+            logger.exception("NGFW creation failed: user=%s name=%s", user.pk, safe_log_value(payload.get("name", "")))
+            return self.bad_request(classify_user_message(str(exc), default="NGFW could not be created"))
+
+        logger.info("NGFW provisioning started: user=%s app_id=%s", safe_log_value(user.email), ngfw_ref.app_id)
+        return Response({"id": str(ngfw_ref.app_id), "name": payload["name"], "status": "provisioning"}, status=201)
+
+
+class NGFWListView(MissionControlReadAPIView):
+    """List user's NGFWs."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _ngfw_read_permission()]
+
+    def get(self, request: Request) -> Response:
+        """Return NGFWs owned by the authenticated actor."""
+        ngfws = cms_list_ngfws(self.actor_user())
+        return Response(
+            {
+                "ngfws": [
+                    {
+                        "id": str(n.app_id),
+                        "name": n.name,
+                        "status": n.status,
+                        "created_at": n.created_at.isoformat(),
+                        "serial_number": n.serial_number,
+                    }
+                    for n in ngfws
+                ]
+            }
+        )
+
+
+class NGFWDestroyView(MissionControlAPIView):
+    """Destroy an NGFW."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _ngfw_write_permission()]
+
+    def post(self, request: Request, app_id: str) -> Response | JsonResponse:
+        """Start NGFW deprovisioning for the requested app id."""
+        data, error = _validated(self, NGFWDestroySerializer, request.data)
+        if error is not None:
+            return error
+        assert data is not None
+        user = self.actor_user()
+        try:
+            _run_ngfw_destroy(user, app_id, data.get("confirm_name", ""))
+        except Http404:
+            raise
+        except _NgfwError as err:
+            return err.response
+
+        logger.info(
+            "NGFW deprovisioning started: user=%s app_id=%s",
+            safe_log_value(user.email),
+            safe_log_value(app_id),
+        )
+        return Response({"status": "deprovisioning"})
+
+
+class CredentialCreateView(MissionControlAPIView):
+    """Create a credential."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _credentials_write_permission()]
+
+    def post(self, request: Request) -> Response | JsonResponse:
+        """Validate and persist a Mission Control credential."""
+        data, error = _validated(self, CredentialCreateSerializer, request.data)
+        if error is not None:
+            return error
+        assert data is not None
+
+        user = self.actor_user()
+        credential_type_slug = data["credential_type"]
+        payload = dict(data)
+        payload["user_id"] = user.id
+        if payload.get("expires_at") == "":
+            payload["expires_at"] = None
+        try:
+            spec = _validate_credential_spec(payload, credential_type_slug)
+            kwargs = spec.model_dump(exclude={"user_id"})
+            cred_ref = _persist_credential(user, credential_type_slug, kwargs)
+        except _CredentialError as err:
+            return err.response
+
+        logger.info(
+            "Credential created: user=%s credential_id=%s type=%s",
+            safe_log_value(user.email),
+            cred_ref.credential_id,
+            safe_log_value(credential_type_slug),
+        )
+        return Response({"id": cred_ref.credential_id, "name": spec.name, "credential_type": credential_type_slug}, 201)
+
+
+class CredentialDeleteView(MissionControlAPIView):
+    """Soft-delete a credential."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _credentials_write_permission()]
+
+    def post(self, request: Request, credential_id: int) -> Response:
+        """Delete a credential visible to the authenticated actor."""
+        user = self.actor_user()
+        try:
+            cms_delete_credential(user, credential_id)
+        except CMSError:
+            raise Http404("Credential not found") from None
+
+        logger.info(
+            "Credential deleted: user=%s credential_id=%s",
+            safe_log_value(user.email),
+            safe_log_value(credential_id),
+        )
+        return Response({"success": True})
+
+
+class ScriptListView(MissionControlReadAPIView):
+    """List experiment scripts for the authenticated user."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _script_read_permission()]
+
+    def get(self, request: Request) -> Response:
+        """Return experiment scripts visible to the authenticated actor."""
+        scripts = list_scripts(self.actor_user())
+        return Response({"scripts": [{"id": s.pk, "name": s.name, "filename": s.original_filename} for s in scripts]})
+
+
+class ScriptUploadView(MissionControlAPIView):
+    """Initiate or complete an experiment-script upload."""
+
+    permission_classes = [IsAuthenticatedSessionOrApiToken, HasMissionControlActor, _script_write_permission()]
+
+    def post(self, request: Request) -> Response:
+        """Dispatch script upload initiation or completion by request shape."""
+        data, error = _validated(self, ScriptUploadSerializer, request.data)
+        if error is not None:
+            return error
+        assert data is not None
+
+        user = self.actor_user()
+        upload_token = data.get("upload_token")
+        if upload_token:
+            return self._complete_script(user, upload_token)
+        return self._initiate_script(user, data)
+
+    def _complete_script(self, user: User, upload_token: str) -> Response:
+        """Complete a pending script upload."""
+        try:
+            script = complete_script_upload(user, upload_token)
+        except ScriptUploadError as exc:
+            logger.exception("Script upload completion failed: user=%s", user.pk)
+            return self.bad_request(classify_user_message(str(exc), default="Upload could not be completed"))
+
+        logger.info("Script upload completed: user=%s script_id=%s", safe_log_value(user.email), script.pk)
+        return Response(
+            {"success": True, "script_id": script.pk, "message": f"Script '{script.name}' uploaded successfully."}
+        )
+
+    def _initiate_script(self, user: User, data: dict[str, Any]) -> Response:
+        """Create a presigned upload destination for an experiment script."""
+        filename = os.path.basename(data["filename"])
+        try:
+            result = initiate_script_upload(user, data["name"], filename, data["file_size"])
+        except ScriptUploadError as exc:
+            logger.exception("Script upload initiation failed: user=%s", user.pk)
+            return self.bad_request(classify_user_message(str(exc), default="Upload could not be initiated"))
+
+        safe_filename = filename.replace("\r", " ").replace("\n", " ").replace("\t", " ")[:200]
+        safe_email = user.email.replace("\r", " ").replace("\n", " ").replace("\t", " ")[:200]
+        logger.info(
+            "Script upload initiated: user=%s filename=%s size=%d",
+            safe_email,
+            safe_filename,
+            data["file_size"],
+        )
+        return Response(result)
